requirements.txt
PyMuPDF
opencv-python
python-docx

requests 
pillow